Assignment Done
Build a table component to display the transactions in a table nicely which supports pagination/sorting/filtering. You will be given a mock api file (attached with this email), create the component in a way it calls to API and populates the table from the data. You need to implement sorting and filtering in both the component and the mock API.

Requirements
You must use React to implement the component
Implement the component from scratch
Table UI is up to you, you can decide to what extent you want to beautify it